refactor: modularize CLI architecture for better maintainability
- Refactored monolithic bin/zfa.dart (1940 lines) into modular structure
- Created separate command handlers in lib/src/commands/
- Extracted models to lib/src/models/
- Added utility modules: file_utils, logger, string_utils
- Both zfa.dart and zuraffa.dart now use shared lib/src/zfa_cli.dart module
- Removed deprecated bin/cli.dart and lib/src/cli.dart
