--- bin/cli.dart
@@ -1,5 +1,0 @@
-import 'package:zuraffa/src/cli.dart' as cli;
-
-void main(List<String> args) {
-  cli.run(args);
-}
