+++ lib/src/generator/observer_generator.dart
@@ -1,0 +1,66 @@
+import 'package:path/path.dart' as path;
+import '../models/generator_config.dart';
+import '../models/generated_file.dart';
+import '../utils/file_utils.dart';
+
+class ObserverGenerator {
+  final GeneratorConfig config;
+  final String outputDir;
+  final bool dryRun;
+  final bool force;
+  final bool verbose;
+
+  ObserverGenerator({
+    required this.config,
+    required this.outputDir,
+    this.dryRun = false,
+    this.force = false,
+    this.verbose = false,
+  });
+
+  Future<GeneratedFile> generate() async {
+    final entityName = config.name;
+    final entitySnake = config.nameSnake;
+    final observerName = '${entityName}Observer';
+    final fileName = '${entitySnake}_observer.dart';
+    final filePath =
+        path.join(outputDir, 'domain', 'usecases', entitySnake, fileName);
+
+    final content = '''
+// Generated by zfa
+
+import 'package:zuraffa/zuraffa.dart';
+import '../../entities/$entitySnake/$entitySnake.dart';
+
+class $observerName extends Observer<$entityName> {
+  final void Function($entityName) onNext;
+  final void Function(AppFailure) onError;
+  final void Function() onComplete;
+
+  $observerName({
+    required this.onNext,
+    required this.onError,
+    required this.onComplete,
+  });
+
+  @override
+  void onNextValue($entityName value) => onNext(value);
+
+  @override
+  void onFailure(AppFailure failure) => onError(failure);
+
+  @override
+  void onDone() => onComplete();
+}
+''';
+
+    return FileUtils.writeFile(
+      filePath,
+      content,
+      'observer',
+      force: force,
+      dryRun: dryRun,
+      verbose: verbose,
+    );
+  }
+}
