+++ lib/src/generator/repository_generator.dart
@@ -1,0 +1,76 @@
+import 'package:path/path.dart' as path;
+import '../models/generator_config.dart';
+import '../models/generated_file.dart';
+import '../utils/file_utils.dart';
+
+class RepositoryGenerator {
+  final GeneratorConfig config;
+  final String outputDir;
+  final bool dryRun;
+  final bool force;
+  final bool verbose;
+
+  RepositoryGenerator({
+    required this.config,
+    required this.outputDir,
+    this.dryRun = false,
+    this.force = false,
+    this.verbose = false,
+  });
+
+  Future<GeneratedFile> generate() async {
+    final repoName = '${config.name}Repository';
+    final fileName = '${config.nameSnake}_repository.dart';
+    final filePath = path.join(outputDir, 'domain', 'repositories', fileName);
+
+    final methods = <String>[];
+
+    for (final method in config.methods) {
+      switch (method) {
+        case 'get':
+          methods.add('  Future<${config.name}> get(${config.idType} id);');
+          break;
+        case 'getList':
+          methods.add('  Future<List<${config.name}>> getList();');
+          break;
+        case 'create':
+          methods.add(
+              '  Future<${config.name}> create(${config.name} ${config.nameCamel});');
+          break;
+        case 'update':
+          methods.add(
+              '  Future<${config.name}> update(${config.name} ${config.nameCamel});');
+          break;
+        case 'delete':
+          methods.add('  Future<void> delete(${config.idType} id);');
+          break;
+        case 'watch':
+          methods.add('  Stream<${config.name}> watch(${config.idType}? id);');
+          break;
+        case 'watchList':
+          methods.add('  Stream<List<${config.name}>> watchList();');
+          break;
+      }
+    }
+
+    final content = '''
+// Generated by zfa
+// zfa generate ${config.name} --methods=${config.methods.join(',')} --repository
+
+import '../entities/${config.nameSnake}/${config.nameSnake}.dart';
+
+abstract class $repoName {
+${methods.join('\n')}
+}
+''';
+
+    return FileUtils.writeFile(
+      filePath,
+      content,
+      'repository',
+      force: force,
+      dryRun: dryRun,
+      verbose: verbose,
+    );
+  }
+}
